[dagster-dbt][refactor] Clean up multi-asset creation (#26614)
## Summary & Motivation

This was mostly a passion project, but it preempts some work that needed to get done in this area in the near future anyway.

We have a bunch of complex logic going on here, purely because we are not taking advantage of the features around passing specs into @multi_asset (because this code predates asset specs!)

Crucially, before this change, we were manually creating internal_asset_deps, AssetIns, and AssetOuts independently, alongside juggling a bunch of different dictionaries to keep track of it all.

Now, I've created a single `get_asset_spec` method. Once we're ready, this can essentially be copy-pasted into the DagsterDbtTranslator class once we're ready, but for now this is purely-internal construct.

The only change in the constructed object is that we no longer have control over the generated output names for each key, and instead they are automatically generated using AssetKey.to_python_identifier. This is the cause of the updated tests.

I judge this is totally fine -- the output names are not actually user-facing in this context, it's just framework code that needs to handle this mapping (and this framework code actually gets a bit nicer as well)

## How I Tested These Changes

## Changelog

NOCHANGELOG