fix Azure doc link rendering (#26763)
## Summary & Motivation
The link for Azure Default Credentials and Secret Credentials in the
AzureBlobComputeLogManager class was formatted as a markdown link
instead of in rst format. This was causing the docs and docs beta build
to fail. There was another minor issue in with sling_mac_arm64 not being
added in the autodoc_mock_imports list.

## How I Tested These Changes

![image](https://github.com/user-attachments/assets/61d430c7-f0a6-4f1f-9618-904ad8312de4)