[components] Templating for asset_attributes (#26633)
## Summary & Motivation

This adds a rendering scope to AssetAttributes, allowing its values to depend on the spec that's currently being evaluated. This is useful for setting properties of a spec based off of other properties (e.g. based off of a tag, set an automation condition)

Refactors component_rendering.py to be more class-based

## How I Tested These Changes

## Changelog

NOCHANGELOG