[components] Replace RequiredScope with RenderedModel (#26759)
## Summary & Motivation

This changes up how we manage marking objects as having deferred fields. The result is a lot more terse, and generally easier to work around.

The base class exposes a method to get all of the rendered properties, which can be overridden in scenarios where we want to customize this. For now, our only existing usage can just use this raw properties dictionary directly

## How I Tested These Changes

## Changelog

NOCHANGELOG