[core-api][experimental] AssetsDefinition `specs` (#26681)
## Summary & Motivation

decision: experimental -> public
reason: we are using this pattern heavily internally, and generally have
built up a lot of APIs around making this sort of thing work. we want
people to be using this pattern in other places (e.g. multi_asset), so
it shouldn't be marked as experimental
docs exist: yes (api only, which is ok)

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.