[dagster-powerbi] Move contextual data from DagsterPowerBITranslator to PowerBITranslatorData (#26654)
## Summary & Motivation

This PR implements `PowerBITranslatorData`, a container class containing
a `PowerBIContentData` object (props) and a `PowerBIWorkspaceData`
object (context), to pass contextual data to the translator without
breaking the `get_asset_spec` signature or requiring context to be
passed to the `__init__` method of the `DagsterPowerBITranslator`.

This is PR is an alternative to the proposals mentioned in
https://github.com/dagster-io/dagster/pull/26617.

## How I Tested These Changes

Updated tests with BK

## Changelog

[dagster-powerbi] Type hints in the signature of
`DagsterPowerBITranslator.get_asset_spec` have been updated - the
parameter `data` is now of type `PowerBITranslatorData` instead of
`PowerBIContentData`. Custom Power BI translators should be updated.