Remove slow validation code in get_partition_set_execution_param_data
Summary:
The current implementation here requires fetching every partition key in the set, which can be very slow. We don't appear to use that for anything other than filtering out keys that aren't in the set, which seems like a reasonable thing to require the callsite to check instead.

BK

> Insert changelog entry or delete this section.