Add a "scheduled_execution_timestamp" field to scheudle tick data
Summary:
This paves the way to make schedule retry behavior more like sensor retry behavior. Right now if a schedule fails, it retries for the same time using the same tick, and the timestamp of that tick is synonymous with the scheduled time in the cron schedule. This differs from sensor retries. This divergence in behavior is confusing and leads to issues like schedule failure alerts leading nowhere, since the schedule tick retried and succeeded.

By adding this field, we can eventually change this behavior so that a retried schedule tick keeps the scheduled_execution_timestamp field, but updates the timestamp field and creates a new tick on the timeline.

> Insert changelog entry or delete this section.