Make CodeExample block SSRd for better SEO. (#26829)
## Summary & Motivation

We want to make sure that the CodeExample block is included in the
statically generated HTML for the site so that it's included in SEO.
Currently since we use a `useEffect` we only fetch the content on the
client.

To fix this we need to make sure we have the content on the server. To
this end I added a check to see if we're on the server and if we are
then we synchronously require the content for the codeblock inline
during render and use it. Otherwise, if we're on the client, we import
the content dynamically and rely on Suspense to avoid unmounting the
SSR'd content during the initial render.

## How I Tested These Changes

`yarn build`
`yarn serve`

Looked at the source in `view-source:http://localhost:3002/` and
confirmed the HTML includes the codeblocks.

There's a brief flash of the light theme but confirmed on the
docs-preview site that this is in an existing issue and unrelated to my
approach.


https://github.com/user-attachments/assets/b760a941-1b77-4c46-b693-d0a79dff234e