[dagster-gcp] live tests (#26761)
## Summary & Motivation
Similar to what we introduced for azure, introduce live testing for gcp.
Starting with the compute log manager.

Just like the azure tests, this runs on branches which touch
dagster-gcp, as well as on the oss nightly pipeline.

Most of the files here are a relative copy-paste job from the azure
tests. There is some new functionality, and I've added comments pointing
to where that lives.

## How I Tested These Changes
New tests pass.