[ui] Lint: error instead of warn (#26856)
## Summary & Motivation

A handful of changes in our eslint-config:

- Change some current warnings to errors. It should be fine for these to
be louder.
- Fix tests.
- Update eslint API usage in `missing-graphql-variables-type`.
- Update the test for `missing-graphql-variables-type` so that
queries/mutations always require the `Variables` parameter. I believe
this was the intent in
https://github.com/dagster-io/dagster/pull/24184/files#diff-6c41e2f7b96b1cef1feff12709d27896bdaee5e96ab3f2622d98283cc57ee200,
but the tests still had "valid" cases that should not have been valid
anymore.
- Bump version number, update `CHANGES`.

## How I Tested These Changes

Manually modify a `useQuery` call to drop its `Variables` type
parameter, verify that it is re-added by lint.

`yarn jest` in `eslint-config`.

`make lint` in `dagster-ui`.