CE 602/rag project implementation (#26795)
## Summary & Motivation

Creating a relatively simple use-case of doing a RAG in Dagster. This
project attempts to replicate the ask-ai functionality and scrapes the
docs site and github issues and discussions, adds them to a pinecone
vectordb. Then there is a query asset that takes in some config
information in the form of a query and then returns a response in the
Dagster UI

## How I Tested These Changes

Definitions loaded properly and assets all materialized.

---------

Co-authored-by: Colton Padden <colton@dagsterlabs.com>