[ui] Move usePrefixedCacheKey into its own file (#26901)
## Summary & Motivation

`usePrefixedCacheKey` currently lives in `AppProvider`, so callsites that import it are importing a big component with an increased risk of circular dependencies.

This is a pretty slim function that just needs `AppContext`, which is defined separately from the provider, so we can just move it out into its own file to mitigate the risk.

## How I Tested These Changes

Buildkite.