[components] Use AssetAttributesModel instead of DbtTranslatorParams (#26768)
## Summary & Motivation

This converts to using the AssetAttributesModel inside the translator. This is nice because it reduces duplicated code (e.g. the AssetKey.from_user_string() conversion), and in general having fewer distinct types flying around is good. (note: the `group_name` key was erroneously called just `group` before, this is something that we now no longer need to worry about)

Unfortunately, the translator code is still a mess on account of not having a unified `get_asset_spec()` method, meaning we have to implement a bunch of individual fns. However, once that's implemented, the translator can become as simple as:

```python

class XComponentTranslator(XTranslator):

    def __init__(self, params: AssetAttributes, resolver): ...

    def get_asset_spec(self, x: ...) -> AssetSpec:
        base_spec = super().get_asset_spec(x)
        rendered_props = params.get_rendered_properties(resolver.with_context(x=x))
        return base_spec.replace_attributes(**rendered_props)

```

## How I Tested These Changes

## Changelog

NOCHANGELOG