Doc 302 new etl tutorial - part 1 (#25320)
## Summary & Motivation
I'm a little way into this and would like to get feedback from
@PedramNavid and @cmpadden on the structure and general flow. This isn't
done at this point, but I figure we could collaborate here and iterate
from there.

I made some changes to the reference file to make it more concise
regarding metadata output. The new code example function works great.

Main Questions I have at this point:

1. Where will we put pictures for the new docs site? I want to add some
UI screenshots for more color.
2. How many explanations should I add here? I know on the previous site,
the main complaint was that there was too much going on. I think having
a couple of sentences explaining a concept is fine, with a link to the
main page for that concept/API
3. Should we hide the code examples and have the user reveal them or
leave them as is?


## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.

---------

Signed-off-by: nikki everett <nikki@dagsterlabs.com>
Co-authored-by: Nikki Everett <neverett@users.noreply.github.com>
Co-authored-by: nikki everett <nikki@dagsterlabs.com>