[perf] Skip expensive `PIPELINE_EXPLORER_ROOT_QUERY` query if viewing an asset job and prefer asset rendering is enabled (#26957)
## Summary & Motivation

When viewing an asset job as an asset graph the
`PIPELINE_EXPLORER_ROOT_QUERY` is unused. This is slowing down the
default case unnecessarily. Short circuit to bypass it.

Note that this uses the `useJob` hook which relies on our workspace
fetching abstraction. If the workspace isn't fetched then we go down the
slow path and its effectively a race between the query and the workspace
fetching.

## How I Tested These Changes

app-proxy

rum:
https://app.datadoghq.com/rum/sessions?query=%40view.id%3A8063e1a8-0612-4aae-8992-0c5ff1623c88%20%40type%3Aview&agg_m=count&agg_m_source=base&agg_t=count&cols=&event=AwAAAZRH7cXuh-qITwAAABhBWlJIN2NYdUFBQ2RXOTdVUDVxYWFMc2gAAAAkMDE5NDQ3ZWQtZWM0OC00ZTQxLWJjOWUtNDcyN2VhNzBhMjMwAAAAGQ&fromUser=false&track=rum&viz=stream&from_ts=1736287373547&to_ts=1736373773547&live=true