[components] Add `--[no-]-use-dg-managed-environment` option to `dagster-dg` (#26943)
## Summary & Motivation

Add a global option to disable manage of environment with `uv`
(`--[no-]dg-managed-environment`) to `dagster-dg`. This has a few
effects:

- Prevent any `dg` command from automatically creating a venv in the
root directory and corresponding `uv.lock`.
- Causes all calls to the `dagster-components` CLI (which is used to
retrieve registered components) to use the current python environment
instead of `uv run`.
- Disables the cache.

NOTE: The reason we disable the cache is that caching logic currently
depends on `uv.lock` to generate the cache key. It is possible we will
later introduce support for the cache without `uv` by falling back to an
alternative mechanism to derive a cache key from the python environment.

## How I Tested These Changes

New unit tests.