Fix spec error conditions in loaders (#27030)
## Summary & Motivation
Previously, even when specs weren't getting loaded, we began erroring on
collisions involving them.

This fixes by excluding specs from collision detection if they aren't
getting loaded. This is still a bit hacky - ideally we can specify up
front exactly what we want to load in a more principled manner.

## How I Tested These Changes
Added new test case for the expected behavior - that when include_specs
is not set, we don't error when a duplicate spec exists, and do error
when it does.

## Changelog
- Fixed a bug with load_assets_from_x functions where we began erroring
when a spec and AssetsDefinition had the same key in a given module. We
now only error in this case if include_specs=True.