[components] Remove translator terminology from SlingReplicationCollectionComponent (#26927)
## Summary & Motivation

The translator language was fairly confusing. Went through some iterations but arrived in a place where the translation layer can be overridden in yaml using the `asset_attributes` field. This field will have the domain-specific object in scope, meaning you can use that to set attributes if desired.

At the same time, this leaves open the ability for users to specify a python-based translator to override the default one.

Made the following changes:

1. Because we want to use the key `asset_attributes` to replace `translator`, this conflicts with the global `asset_attributes` field, so I renamed this to be `asset_transforms`. This gives a better sense of the fact that you're doing specific operations over all assets (imo)
2. Removed the weird subclassing stuff with the existing asset transforms -- this was only necessary in a world where merge / replace supported different argument sets, but this is no longer the case so the extra complexity was unnecessary
3. Added a generic wrapper layer to make it easier to create these sorts of translator wrappers that incorporate both a base translator as well as jinja templating

## How I Tested These Changes

## Changelog

NOCHANGELOG