[components] Allow specification of rendering scope (#26952)
## Summary & Motivation

This creates a new FieldInfo subclass, which carries additional metadata about what the resolved field will look like, and what data will be available to resolve it.

We have to create a subclass of FieldInfo because pydantic will only recognize things inside of the `Annotated[]` block if they are instances of FieldInfo, and the base FieldInfo does not have any fields that we can stash arbitrary objects in.


## How I Tested These Changes

## Changelog

NOCHANGELOG