Add a "consecutive_failure_count" field to TickData (#26790)
## Summary & Motivation
Add a field to TickData capturing the number of consecutive times that
this schedule or sensor has failed (across multiple ticks). This differs
from the existing failure_count field, which tracks whether or not a
particular tick should retry or not before moving on. This allows you to
determine when a given sensor has failed more than a certain number of
times consecutively.

## How I Tested These Changes
New test cases

## Changelog
NOCHANGELOG