[ui] Virtualize timeline run popover list (#27062)
## Summary & Motivation

The "+X more" link in the timeline popover is broken for segments for  "Ad hoc materializations" and automations without specific code objects.

Remove the link and virtualize the list instead.

## How I Tested These Changes

View a timeline that has chunks with lots of runs in them. Hover, verify that the virtualized list renders and behaves properly.

Find a chunk with just a couple of runs, verify that the popover height is not excessive.

## Changelog

[ui] In the run timeline, replace the "+ more" link in the hover popover by displaying all run IDs and timings in the list.