fix additional_deps not showing up on assetsdefinition (#27037)
## Summary & Motivation
Previously, `keys_by_input_name` did not contain the ins specified by
additional_ins or additional_deps in the @asset_check decorator.

This fixes to include those keys as part of the property.

This was causing
https://github.com/dagster-io/dagster/issues/27003#issuecomment-2583349296
in some yet to be determined way. We've been able to confirm the fix,
but have not yet pinpointed a straightforward repro.

## How I Tested These Changes
- Fixes the problematic edge case. Add a bunch of new assertions to
prevent regressions in the future.
## Changelog
- Fixed a bug with asset checks where additional_deps/additional_ins
were not being threaded through properly in certain cases, and would
result in errors at job creation.