[ui] Add tabs to evaluation dialog (#26960)
## Summary & Motivation

Update `EvaluationDetailDialog` to have two tabs: Evaluation and Runs. The first shows what the dialog currently shows, which is the details of the evaluation itself. The second is a filtered Runs table with the runs associated with the evaluation.

<img width="1259" alt="Screenshot 2025-01-14 at 10 52 53" src="https://github.com/user-attachments/assets/8e075249-38ad-417d-9fa4-0dd6c2661d4e" />

<img width="1252" alt="Screenshot 2025-01-14 at 10 52 48" src="https://github.com/user-attachments/assets/76858c12-d840-4ead-841f-065c35ddec4c" />


## How I Tested These Changes

View dialog from automation sensor and asset automations list. Verify that the tabs behave and render correctly.

## Changelog

[ui] When viewing details of an automation evaluation, show the relevant runs in a tab within the same dialog.