[components][docs] Schema viewer (#26978)
## Summary & Motivation

Extremely bare-bones sample yaml schema generator.

Currently this spits out a "maximal" version of the schema, including all optional parameters, which is better suited for docs usecases.

If we adapt this to be used in the component generation sequence, we'll probably want to use a minimal example.

The one thing here that adds an unfortunate amount of complexity is the injection of comments into the generated yaml (particularly to denote the available scope). Yaml libraries don't handle comments very nicely and so we need to do some fancy stuff to get this to work.


![Screenshot 2025-01-09 at 2.45.14 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/cTPQ7oSxo9qxXgkzpnaK/801959df-41bf-4b4b-b81f-22b15f212dd8.png)

## How I Tested These Changes

## Changelog

NOCHANGELOG