[components][fix] Fix issue with additional scope determination (#27043)
## Summary & Motivation

Fixes a weird issue I ran into. In essence, in pydantic 3.8.x, the generated json schema will have the `json_schema_extras` stored on dereferenced json schemas of any subtype, whereas in 3.10.x, the extras will be stored on the reference itself. No idea why this is different but it is what it is.


## How I Tested These Changes

## Changelog

NOCHANGELOG