[bugfix] Restart k8s log stream on urllib3 failure (#26760)
## Summary & Motivation
Should fix the bug described here -
https://github.com/dagster-io/dagster/issues/26626

The `execute_k8s_job` method uses `watch.stream()` to stream logs from
k8s pods. When the client enters a stale state, we should call `stream`
again. See the bug report for more information.

## How I Tested These Changes
I was unable to fix a repeatable way for recreating the issue and there
are not existing tests for `execute_k8s_job`.
I deployed a similar fix to our dev and prod environments, and the
problem has not appear yet. At the very least I can say that it didn't
degrade the stability of this method.

## Changelog

> Insert changelog entry or delete this section.