[ui] Display asset checks targeted by sensors on the Automations page (#26971)
## Summary & Motivation

This PR addresses
https://linear.app/dagster-labs/issue/FE-714/fix-empty-dialog-issue-in-target-dialog-on-sensors-page

Some sensors target asset check selections but the targeted checks were
not visible when you clicked on the tag for more information. We now
fetch the checks and render them in the target modal. I also unified the
visual appearance of the asset keys + asset checks in this modals by
reusing utils we wrote for the new runs feed targets column.

Sidenote: I'm talking with Josh about a small follow-up PR to address
the appearance of these tags, which are clipped very often. Will stack
another PR on this one to fix.

## How I Tested These Changes

I tested these manually using the root workspace, which has examples of
all three types of content that can appear in this modal.

![Screenshot 2025-01-09 at 9 17
47 AM](https://github.com/user-attachments/assets/22b65517-395d-442d-bbcd-01ae23d8b5ac)
![Screenshot 2025-01-09 at 9 16
49 AM](https://github.com/user-attachments/assets/25e4fd21-a043-469a-bbd1-43198678a022)

## Changelog

[ui] Sensors targeting asset checks now list the asset checks when you
click to view their targets.

---------

Co-authored-by: bengotow <bgotow@elementl.com>