time_window_partitions: add get_partition_key helper method that converts date/datetime to str partition key (#26511)
## Summary & Motivation

Create a helper method `get_partition_key` on
`TimeWindowPartitionsDefinition` that allows to get a string
representation of a partition key using a string, datetime or date. This
can be convenient for hourly/daily/weekly/monthly partitioned assets, as
the method will automatically use the relevant partition definition
format.

The current implementation will error if it gets a date/datetime that is
not a valid partition key for the given partition definition.

## How I Tested These Changes

Added tests in `test_partitioned_assets.py`. Also, tested in a local
Dagster.

## Changelog

> Add get_partition_key helper method that can be used on
hourly/daily/weekly/montly partitioned assets to get partition key using
date or datetime.