[dagster-azure]: Add Azure specific information to captured logs (#27090)
## Summary & Motivation

External cloud storage based compute log managers are often configured to only collect and show the url. However, in the case of the AzureBlobComputeLogManager, the provided URL will not be clickable if the storage account or container are private, which is a reasonable choice for commercial users using this feature for security and compliance on Dagster+.

This supports these users by providing both a more easily readable path and a clickable shell command to retrieve the log locally.

How this is implemented is by adding new fields to CapturedLogData to pass the path as well as extra azure specific information so that the frontend can render the shell cmd.

## Note to reviewers

This is a rework of #26982 - applying the changes requested and discussed. 

## How I Tested These Changes
- dagster dev local (with the AzureBlobComputeLogManager as well as the default log manager to ensure other code paths aren't broken).
- added testing of new fields in integration tests

## Changelog

- [dagster-azure] AzureBlobComputeLogManager provides shell command to download logs.