feat(dagster_polars): added support for rust engine, fixed minor typos (#27107)
## Summary & Motivation
I was working with `dagster_polars` to write a delta lake table and
noticed that it fails in recent versions of `deltalake` because [of this
line](https://github.com/delta-io/delta-rs/blob/0cfde9656bde2219f52ea5ac4662bf18449ed739/python/deltalake/writer.py#L310).
The problem can be worked around by setting `{"engine": "pyarrow"}` in
the asset metadata, however they say that this engine will be removed on
version 1.0, so instead here I added a method to generate the
`predicate` needed by the rust engine.

Because in tests (or at least in the development environment)
`deltalake==0.17.3` is installed this behavior was not noticed.

Referenced documentation:
https://delta-io.github.io/delta-rs/usage/writing/#overwriting-part-of-the-table-data-using-a-predicate

I also fixed several typos in the same file.

## How I Tested These Changes
Parametrized existing tests to cover both engines.

## Changelog
- Added `rust` engine support for `dagster_polars[deltalake]` when
native partitioning