Add `.allow()` and `.ignore()` for bool conditions (#27147)
## Summary & Motivation

Users often want to make small changes to the existing built-in
policies. However, these utilities aren't always the most ergonomic for
certain use cases. For example, dep conditions support allowing or
ignoring an asset selection, but this has to be done individually for
each sub-condition of a complex condition.

## How I Tested These Changes

Added unit tests.

## Changelog

Implement the `.allow()` and `.ignore()` methods on boolean conditions,
which will propagate the asset selection down to any contained dep
sub-conditions.