[1/n][dagster-dlt] Move asset spec creation to DagsterDltTranslator.get_asset_spec (#27207)
## Summary & Motivation

Move the asset spec creation from the asset decorator to a
`get_asset_spec` method in the translator.

Also updates adds tests. No additional test was added for group names,
owners, automation condition and auto-materialize policies. It is not
recommended to update these values by overriding `get_asset_spec`
because they are not related to dlt specifically.

## How I Tested These Changes

Updated and new tests with BK