[5/n] Refactor SelectionInput auto-complete logic (#27229)
## Summary & Motivation

Couple goals:
- Separate the selection visitor into a Base visitor which encapsulates
the cursor based visiting logic, and the SelectionVisitor which makes
auto-complete suggestions. The reason is that cloud will be using the
base visitor to figure out which auto-completions we need to query for.
- Make `SelectionAutoCompleteInput` take a `useAutoComplete` hook as a
prop which can return `loading: true` to indicate asynchronous
suggestions are in progress. This will be used by cloud for elastic
search filtering.

## How I Tested These Changes

1. Existing jests
2. New jest tests for BaseVisitor
3. Used this in cloud for elastic search filtering