[docs] add docs for PipesEMRContainersClient (#27159)
## Summary & Motivation

This PR adds docs for the new PipesEMRContainersCient. 

It sounds confusing but AWS EMR Containers and AWS EMR on EKS is
actually the same thing.
The former is the name of the AWS API, and the latter is the service
name (sounds like the same thing huh?) which is more used in human
language rather than automation.

For example, the `boto3` client is called `emr-containers`. 

I tried to preserve this differentiation in these docs.

## How I Tested These Changes

The snippets were tested with a real EMR on EKS cluster.