More aggressively mask user code errors when masking enabled (#27183)
## Summary

Follow-on to the functionality introduced in #18688. The original PR
ensures that if the `DAGSTER_REDACT_USER_CODE_ERRORS` env var is set to
`1`, instances of `DagsterUserCodeExecutionError` which are serialized
are masked so that their potentially sensitive contents are not leaked.
However, there is a class of errors that can still expose elements of
user code: system/framework errors or interrupts, for example, will
include their stack trace and snippets of user code.

This PR attempts to handle this second class of errors, in a few ways:
- Any error which is raised in a user code context has its stacktrace
logged and then cleared, so if the error makes it to serialization, it
will only include the error message.
- User code exceptions have their `original_exc_info` purged.
- A global mapping stores the error, so that a unique error ID can be
raised in the UI, which can be used to locate the original error in
logs.

## Test Plan

New unit tests which specifically handle the intterrupt/system error
cases.