Fix MultiPartitionsDefinition subset -> range -> keys conversion inconsistency (#26652)
## Summary & Motivation
This bug reported a backfill of a multi-partitioned asset never
completing https://github.com/dagster-io/dagster/issues/26307. I was
able to replicate this issue and believe the cause is that we are losing
information when we convert a sequence of MultiParitition keys to a
subset, then to a range, then back to a list of keys. I've written up a
test case demonstrating the behavior.

It looks like this conversion used to work until
https://github.com/dagster-io/dagster/pull/19448. This PR updated
`MultiPartitionsDefinition.get_partition_keys_in_range` but not the
other conversion methods

This PR fixes this issue by fixing how we construct ranges from a subset
for MultiPartitions. Rather than iterating through the flat list of all
partition keys, we break the partition keys into groups based on the
primary key. Then we construct the range. For example if we have
partition dimensions `A, B, C` and `1, 2, 3, 4` we can visualize the
partitions keys as

```
       1   2   3   4
A | A1 A2 A3 A4
B | B1 B2 B3 B4
C | C1 C2 C3 C4
```

To construct the range for a subset we would consider the partition keys
in chunks `A1 A2 A3 A4`, `B1 B2 B3 B4`, and `C1 C2 C3 C4` and construct
the range for each chunk. This is not the most efficient "packing" of
keys into ranges, but fixes the problem of dropping partition keys.

---

documenting how this manifests as an issue for backfills

In backfills this manifests like this:
- We have a list of MultiPartition keys to materialize for an asset:
`2023-01-01|a, 2023-01-01|b, 2023-01-01|c, 2023-01-02|a`
- We convert those keys into a subset, and then generate the ranges to
determine which partitions can be materialized together in a single run.
The `get_partition_key_ranges` returns the range `
[PartitionKeyRange(start='2023-01-01|a', end='2023-01-02|a')]`. This is
accurate since if you list the partition keys in order, the sequence
goes `2023-01-01|a, 2023-01-01|b, 2023-01-01|c, 2023-01-02|a,
2023-01-02|b, ...`
- We launch a single run for the range
[PartitionKeyRange(start='2023-01-01|a', end='2023-01-02|a')]`
- On the next tick of the backfill daemon, we want to see which
partitions were materialized. When we get the materialization events for
the asset, we only see events for partitions `2023-01-01|a` and
`2023-01-02|a`. I believe this is because we are calling
`get_partition_keys_in_range` when determining what materialization
events to emit for the ranged run (still tracking down where this
happens in the code)
- There's an additional layer of logic in the backfill daemon that won't
see that the runs for `2023-01-01|b, 2023-01-01|c` because they are root
assets for the backfill, and we could make some changes to see that they
didn't get materialized. But the core bug is with these conversion
methods


## How I Tested These Changes
New unit tests, these tests fail in master

## Changelog

[bug fix] Fixed an issue where converting a list of partition keys for a
`MultiPartitionsDefinition` to a `PartitionSubset`, then computed the
ranges for the subset, then converted back to a list of partition keys,
some partition keys would be dropped.