[1/n][dagster-looker] Move asset spec creation to DagsterLookerLkmlTranslator.get_asset_spec (#27249)
## Summary & Motivation

Move the asset spec creation from the asset decorator to a
`get_asset_spec` method in the translator.

Also updates adds tests. No additional test was added for group names,
owners and tags. It is not recommended to update these values by
overriding `get_asset_spec` because they are not related to Looker
specifically.

## How I Tested These Changes

Updated and new tests with BK